Completely refactoring in progress database utils
There were too many bugs and it was getting confusing to have to figure
out how to get everything from everywhere, so I'm limiting the access so
that you go through the util functions to get the value. Quite a bit
more code, but should also be quite a bit easier to follow. Part 1 in
refactor series